refactor(sandbox): relocate docs/playground → docs/sandbox; merge develop
develop retired the playground scaffold (PR #22: removed docs/playground/component
+ packages/playground). This branch's sandbox build extended exactly those files,
so it is relocated to a self-contained docs/sandbox/ home (carrying the runner.ts
contract + StitchPlayground.tsx that develop deleted) and merged up to current
develop (9d49888). docs/playground/ now holds only develop's surviving spec docs,
untouched. All cross-package imports repointed to docs/sandbox/contracts.

Verified post-move: 10/10 suites green; contracts/runtime/sandbox-sim tsc 0.
