+++ docs/sandbox/component/StitchPlayground.tsx
