+++ docs/sandbox/component/output-format.ts
