+++ docs/sandbox/runtime/shims/otlp-browser.ts
