+++ docs/sandbox/runtime/vendor.d.ts
