+++ packages/sandbox-sim/src/adapters/node.ts
@@ -17,7 +17,7 @@
 import type {
     SimHandler,
     SimRequest,
-} from '../../../../docs/playground/contracts/sim';
+} from '../../../../docs/sandbox/contracts/sim';
 import { dispatch } from '../dispatch';
 
 /** Build a SimRequest from the raw fetch arguments. */
