+++ packages/sandbox-sim/src/dispatch.test.ts
@@ -9,7 +9,7 @@ import type {
     SimKnobs,
     SimRequest,
     SimResponse,
-} from '../../../docs/playground/contracts/sim';
+} from '../../../docs/sandbox/contracts/sim';
 import { createFetchShim } from './adapters/node';
 import { dispatch, resetFlaky } from './dispatch';
