+++ packages/sandbox-sim/src/dispatch.ts
@@ -18,7 +18,7 @@ import type {
     SimKnobs,
     SimRequest,
     SimResponse,
-} from '../../../docs/playground/contracts/sim';
+} from '../../../docs/sandbox/contracts/sim';
 
 // ---------------------------------------------------------------------------
 // Sandbox-404 helper (SANDBOX.md §4.3)
