+++ packages/sandbox-sim/src/handlers/auth-resilience.ts
@@ -31,7 +31,7 @@ import type {
     SimKnobs,
     SimRequest,
     SimResponse,
-} from '../../../../docs/playground/contracts/sim';
+} from '../../../../docs/sandbox/contracts/sim';
 
 // ---------------------------------------------------------------------------
 // GET /auth/me
